+++ requirements.txt
@@ -1,0 +1,3 @@
+Flask
+pygame
+Pillow
