+++ templates/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Black Metal Match Game</title>
+</head>
+<body>
+    <h1>Black Metal Match-3 Game</h1>
+    <p>The game is running locally. Check the game window.</p>
+</body>
+</html>
